fix macos build

+++ compiler/compiler-settings.h
@@ -161,7 +161,7 @@ public:
   KphpOption<bool> dynamic_incremental_linkage;
 
   KphpOption<double> debug_info_force_disable_prob;
-  KphpOption<size_t> debug_info_force_disable_prob_seed;
+  KphpOption<uint64_t> debug_info_force_disable_prob_seed;
 
   KphpOption<uint64_t> profiler_level;
   KphpOption<bool> enable_global_vars_memory_stats;
